UOS
#include <iostream>
#include <sys/stat.h>
#include <cstdio>

using namespace std;
These lines include the necessary C++ standard library and system header files for the program, and also bring the std namespace into scope. 

int main(int argc, char *argv[]) {
    if (argc != 2) {
        cerr << "Usage: " << argv[0] << " <filename>" << endl;
        return 1;
    }
This is the main function that starts the program execution. It takes two arguments: argc, the number of command-line arguments passed to the program, and argv, an array of strings containing those arguments. The if statement checks if the number of arguments is not equal to two (i.e., the program was not called with a filename argument). If so, it prints an error message to the standard error stream (cerr) and exits the program with a non-zero status code.

c++
Copy code
    struct stat s;
    if (stat(argv[1], &s) != 0) {
        perror("stat");
        return 1;
    }
This creates a struct stat variable s to hold the file information, and calls the stat() function to retrieve information about the file specified by the first command-line argument (argv[1]). If stat() returns a non-zero value, it means there was an error and the program prints an error message using the perror() function, which prints a descriptive error message to stderr. The program then exits with a non-zero status code.

c++
Copy code
    cout << "Inode: " << s.st_ino << endl;
    cout << "Mode: " << s.st_mode << endl;
    cout << "Links: " << s.st_nlink << endl;
    cout << "UID: " << s.st_uid << endl;
    cout << "GID: " << s.st_gid << endl;
    cout << "Size: " << s.st_size << endl;
    cout << "Block size: " << s.st_blksize << endl;
    cout << "Blocks: " << s.st_blocks << endl;
If stat() succeeds, the program prints various information about the file to the standard output stream (cout), including the inode number, file mode, number of hard links, owner user ID and group ID, file size, block size, and number of blocks used by the file.

c++
Copy code
    return 0;
}
This line indicates a successful exit status code, and terminates the program execution.